Moved model's selection attribute methods from LiveSelection to Selection.
